LVRUBYM-221:Removed version for services
