+++ db/schema.rb
@@ -1,6 +1,5 @@
 ActiveRecord::Schema.define(version: 20170413145257) do
 
-  # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
   create_table "groups", force: :cascade do |t|
